chore: apply code reviews
test: fix failing tests
chore: implement method that was not
chore: remove unused utility function
