+++ apps/editor/src/__test__/unit/markdown/util.ts
@@ -22,18 +22,6 @@ export function removeDataAttr(html: string) {
   return html.replace(/\sdata-nodeid="\d{1,}"/g, '').trim();
 }
 
-export function removeProseMirrorHackNodes(html: string) {
-  const reProseMirrorImage = /<img class="ProseMirror-separator" alt="">/g;
-  const reProseMirrorTrailingBreak = / class="ProseMirror-trailingBreak" \//g;
-
-  let resultHTML = html;
-
-  resultHTML = resultHTML.replace(reProseMirrorImage, '');
-  resultHTML = resultHTML.replace(reProseMirrorTrailingBreak, '');
-
-  return resultHTML;
-}
-
 export function createHTMLrenderer() {
   const customHTMLRenderer: HTMLConvertorMap = {
     htmlBlock: {
