+++ plugins/table-merged-cell/src/__test__/integration/wysiwyg/helper/cellSelection.ts
@@ -33,7 +33,6 @@ function createTableFragment(tableHead: Node, tableBody: Node) {
   return Fragment.from(fragment);
 }
 
-// @ts-ignore
 export default class CellSelection extends Selection {
   private offsetMap: TableOffsetMap;
 
@@ -131,5 +130,9 @@ export default class CellSelection extends Selection {
       targetNode.content = targetNode.content.append(Fragment.from(copiedRow));
     }
     return new Slice(createTableFragment(tableHead, tableBody), 1, 1);
+  }
+
+  toJSON() {
+    return JSON.stringify(this);
   }
 }
